implemented handleNewCarSubmit function in newCarModal.jsx and fixed type error for odometer type in typeDefs.js

+++ server/schemas/typeDefs.js
@@ -53,7 +53,7 @@ const typeDefs = gql`
       model: String!
       year: Int!
       raceNumber: Int
-      odometer: Float
+      odometer: Int
     ): Car
   }
 `;
